txscript: Export several useful funcs for treasury.
This PR moves and exports several functions that are independently
useful.

This is one of several PRs that will follow in order to make the
treasury PR a bit smaller.
