+++ txscript/opcode.go
@@ -2465,13 +2465,13 @@ func opcodeCheckSig(op *opcode, data []byte, vm *Engine) error {
 	// the data stack.
 	hashType := SigHashType(fullSigBytes[len(fullSigBytes)-1])
 	sigBytes := fullSigBytes[:len(fullSigBytes)-1]
-	if err := vm.checkHashTypeEncoding(hashType); err != nil {
+	if err := CheckHashTypeEncoding(hashType); err != nil {
 		return err
 	}
-	if err := vm.checkSignatureEncoding(sigBytes); err != nil {
+	if err := CheckSignatureEncoding(sigBytes); err != nil {
 		return err
 	}
-	if err := vm.checkPubKeyEncoding(pkBytes); err != nil {
+	if err := CheckPubKeyEncoding(pkBytes); err != nil {
 		return err
 	}
 
@@ -2662,10 +2662,10 @@ func opcodeCheckMultiSig(op *opcode, data []byte, vm *Engine) error {
 		// Only parse and check the signature encoding once.
 		var parsedSig *ecdsa.Signature
 		if !sigInfo.parsed {
-			if err := vm.checkHashTypeEncoding(hashType); err != nil {
+			if err := CheckHashTypeEncoding(hashType); err != nil {
 				return err
 			}
-			if err := vm.checkSignatureEncoding(signature); err != nil {
+			if err := CheckSignatureEncoding(signature); err != nil {
 				return err
 			}
 
@@ -2687,7 +2687,7 @@ func opcodeCheckMultiSig(op *opcode, data []byte, vm *Engine) error {
 			parsedSig = sigInfo.parsedSignature
 		}
 
-		if err := vm.checkPubKeyEncoding(pubKey); err != nil {
+		if err := CheckPubKeyEncoding(pubKey); err != nil {
 			return err
 		}
 
@@ -2834,7 +2834,7 @@ func opcodeCheckSigAlt(op *opcode, data []byte, vm *Engine) error {
 	// requirements enabled by the flags.
 	hashType := SigHashType(fullSigBytes[len(fullSigBytes)-1])
 	sigBytes := fullSigBytes[:len(fullSigBytes)-1]
-	if err := vm.checkHashTypeEncoding(hashType); err != nil {
+	if err := CheckHashTypeEncoding(hashType); err != nil {
 		return err
 	}
